+++ src/common/defaultSetting.js
@@ -3,7 +3,7 @@ const os = require('os')
 const { isMac } = require('./utils')
 
 const defaultSetting = {
-  version: '1.0.34',
+  version: '1.0.38',
   player: {
     togglePlayMethod: 'listLoop',
     highQuality: false,
@@ -12,6 +12,7 @@ const defaultSetting = {
     isMute: false,
     mediaDeviceId: 'default',
     isMediaDeviceRemovedStopPlay: false,
+    isShowLyricTransition: true,
   },
   desktopLyric: {
     enable: false,
@@ -22,6 +23,7 @@ const defaultSetting = {
     x: null,
     y: null,
     theme: 0,
+    isLockScreen: true,
     style: {
       fontSize: 120,
       opacity: 95,
@@ -35,7 +37,7 @@ const defaultSetting = {
     isSaveScrollLocation: true,
   },
   download: {
-    enable: false,
+    enable: true,
     savePath: path.join(os.homedir(), 'Desktop'),
     fileName: '歌名 - 歌手',
     maxDownloadNum: 3,
@@ -48,7 +50,7 @@ const defaultSetting = {
     tabId: 'kw__16',
   },
   songList: {
-    source: 'kg',
+    source: 'kw',
     sortId: '5',
     tagInfo: {
       name: '默认',
@@ -81,11 +83,12 @@ const defaultSetting = {
     themeId: 0,
   },
   windowSizeId: 2,
-  themeId: 0,
+  themeId: 6,
   langId: null,
   sourceId: 'kw',
-  apiSource: 'temp',
+  apiSource: 'yj',
   sourceNameType: 'alias',
+  isShowAnimation: true,
   randomAnimate: true,
   ignoreVersion: null,
   isAgreePact: false,