+++ src/main/modules/winLyric/event.js
@@ -1,16 +1,19 @@
 const { common: COMMON_EVENT_NAME, winLyric: WIN_LYRIC_EVENT_NAME, hotKey: HOT_KEY_EVENT_NAME, mainWindow: MAIN_WINDOW_EVENT_NAME } = require('../../events/_name')
 const { mainSend, NAMES: { winLyric: ipcWinLyricNames } } = require('../../../common/ipc')
 const { desktop_lyric } = require('../../../common/hotKey')
+const { getLyricWindowBounds } = require('./utils')
 
 let isLock = null
 let isEnable = null
 let isAlwaysOnTop = null
+let isLockScreen = null
 const setLrcConfig = () => {
   let desktopLyric = global.appSetting.desktopLyric
   if (global.modules.lyricWindow) {
     mainSend(global.modules.lyricWindow, ipcWinLyricNames.set_lyric_config, {
       config: desktopLyric,
       languageId: global.appSetting.langId,
+      isShowLyricTransition: global.appSetting.player.isShowLyricTransition,
     })
     if (isLock != desktopLyric.isLock) {
       isLock = desktopLyric.isLock
@@ -24,6 +27,17 @@ const setLrcConfig = () => {
       isAlwaysOnTop = desktopLyric.isAlwaysOnTop
       global.modules.lyricWindow.setAlwaysOnTop(desktopLyric.isAlwaysOnTop, 'screen-saver')
     }
+    if (isLockScreen != desktopLyric.isLockScreen) {
+      isLockScreen = desktopLyric.isLockScreen
+      if (desktopLyric.isLockScreen) {
+        global.modules.lyricWindow.setBounds(getLyricWindowBounds(global.modules.lyricWindow.getBounds(), {
+          x: null,
+          y: null,
+          w: desktopLyric.width,
+          h: desktopLyric.height,
+        }))
+      }
+    }
   }
   if (isEnable != desktopLyric.enable) {
     isEnable = desktopLyric.enable
@@ -36,7 +50,7 @@ const setLrcConfig = () => {
 }
 global.lx_event.common.on(COMMON_EVENT_NAME.config, name => {
   if (WIN_LYRIC_EVENT_NAME.name === name) return
-  setLrcConfig(false)
+  setLrcConfig()
 })
 
 global.lx_event.mainWindow.on(MAIN_WINDOW_EVENT_NAME.setLyricInfo, info => {