+++ src/renderer/components/core/PlayerDetail.vue
@@ -4,35 +4,47 @@
     //- div(:class="$style.bg2")
     div(:class="$style.header")
       div(:class="$style.controBtn")
-        button(type="button" :class="$style.hide" :title="$t('core.player.hide_detail')" @click="visiblePlayerDetail(false)")
+        button(type="button" :class="$style.hide" :tips="$t('core.player.hide_detail')" @click="visiblePlayerDetail(false)")
           svg(:class="$style.controBtnIcon" version='1.1' xmlns='http://www.w3.org/2000/svg' xlink='http://www.w3.org/1999/xlink' width='80%' viewBox='0 0 30.727 30.727' space='preserve')
             use(xlink:href='#icon-window-hide')
-        button(type="button" :class="$style.min" :title="$t('core.toolbar.min')" @click="min")
+        button(type="button" :class="$style.min" :tips="$t('core.toolbar.min')" @click="min")
           svg(:class="$style.controBtnIcon" version='1.1' xmlns='http://www.w3.org/2000/svg' xlink='http://www.w3.org/1999/xlink' width='100%' viewBox='0 0 24 24' space='preserve')
             use(xlink:href='#icon-window-minimize')
 
         //- button(type="button" :class="$style.max" @click="max")
-        button(type="button" :class="$style.close" :title="$t('core.toolbar.close')" @click="close")
+        button(type="button" :class="$style.close" :tips="$t('core.toolbar.close')" @click="close")
           svg(:class="$style.controBtnIcon" version='1.1' xmlns='http://www.w3.org/2000/svg' xlink='http://www.w3.org/1999/xlink' width='100%' viewBox='0 0 24 24' space='preserve')
             use(xlink:href='#icon-window-close')
 
-    div(:class="$style.main")
+    div(:class="[$style.main, isShowComment ? $style.showComment : null]")
       div(:class="$style.left")
+        //- div(:class="$style.info")
         div(:class="$style.info")
-          div(:class="$style.img")
-            img(:src="musicInfo.img" v-if="musicInfo.img")
-          div(:class="$style.description")
+          img(:class="$style.img" :src="musicInfo.img" v-if="musicInfo.img")
+          div(:class="$style.description" @click="isShowComment = true")
             p {{$t('core.player.name')}}{{musicInfo.name}}
             p {{$t('core.player.singer')}}{{musicInfo.singer}}
             p(v-if="musicInfo.album") {{$t('core.player.album')}}{{musicInfo.album}}
-        //- div(:class="$style.list")
-          ul
 
       div(:class="$style.right")
         div(:class="[$style.lyric, lyricEvent.isMsDown ? $style.draging : null]" @wheel="handleWheel" @mousedown="handleLyricMouseDown" ref="dom_lyric")
           div(:class="$style.lyricSpace")
           p(v-for="(info, index) in lyricLines" :key="index" :class="lyric.line == index ? $style.lrcActive : null") {{info.text}}
           div(:class="$style.lyricSpace")
+
+      div(:class="$style.comment")
+        div(:class="$style.commentHeader" @click="isShowComment = false")
+          h3 {{title}}
+        div.scroll(:class="$style.commentMain")
+          div
+            h2(:class="$style.commentType") 热门评论
+            material-comment-floor(v-if="comment.hotComments.length" :class="$style.reply_floor" :comments="comment.hotComments")
+          div
+            h2(:class="$style.commentType") 最新评论
+            material-comment-floor(v-if="comment.comments.length" :class="$style.reply_floor" :comments="comment.comments")
+          div(:class="$style.pagination")
+            material-pagination(:count="comment.total" :limit="comment.limit" :page="comment.page" @btn-click="handleToggleCommentPage")
+
     div(:class="$style.footer")
       div(:class="$style.left")
         div(:class="$style.progressContainer")
@@ -50,15 +62,15 @@
       div(:class="$style.playControl")
         //- div(:class="$style.playBtn")
         //- div(:class="$style.playBtn")
-        div(:class="$style.playBtn" @click="$emit('action', { type: 'prev' })" style="transform: rotate(180deg);" :title="$t('core.player.prev')")
+        div(:class="$style.playBtn" @click="$emit('action', { type: 'prev' })" style="transform: rotate(180deg);" :tips="$t('core.player.prev')")
           svg(version='1.1' xmlns='http://www.w3.org/2000/svg' xlink='http://www.w3.org/1999/xlink' height='100%' viewBox='0 0 220.847 220.847' space='preserve')
             use(xlink:href='#icon-nextMusic')
-        div(:class="$style.playBtn" :title="isPlay ? $t('core.player.pause') : $t('core.player.play')" @click="$emit('action', { type: 'togglePlay' })")
+        div(:class="$style.playBtn" :tips="isPlay ? $t('core.player.pause') : $t('core.player.play')" @click="$emit('action', { type: 'togglePlay' })")
           svg(v-if="isPlay" version='1.1' xmlns='http://www.w3.org/2000/svg' xlink='http://www.w3.org/1999/xlink' height='100%' viewBox='0 0 277.338 277.338' space='preserve')
             use(xlink:href='#icon-pause')
           svg(v-else version='1.1' xmlns='http://www.w3.org/2000/svg' xlink='http://www.w3.org/1999/xlink' height='100%' viewBox='0 0 170 170' space='preserve')
             use(xlink:href='#icon-play')
-        div(:class="$style.playBtn" @click="$emit('action', { type: 'next' })" :title="$t('core.player.next')")
+        div(:class="$style.playBtn" @click="$emit('action', { type: 'next' })" :tips="$t('core.player.next')")
           svg(version='1.1' xmlns='http://www.w3.org/2000/svg' xlink='http://www.w3.org/1999/xlink' height='100%' viewBox='0 0 220.847 220.847' space='preserve')
             use(xlink:href='#icon-nextMusic')
 </template>
@@ -205,6 +217,109 @@ export default {
       _lyricLines: [],
       lyricLines: [],
       isSetedLines: false,
+      isShowComment: false,
+      comment: {
+        page: 1,
+        total: 10,
+        maxPage: 1,
+        limit: 5,
+        comments: [{
+          text: ['123123hhh'],
+          userName: 'dsads',
+          avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+          time: '2020-10-22 22:14:17',
+          likedCount: 100,
+          reply: [
+            {
+              text: ['123123hhh'],
+              userName: 'dsads',
+              avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+              time: '2020-10-22 22:14:17',
+              likedCount: 100,
+            },
+          ],
+        }, {
+          text: ['123123hhh'],
+          userName: 'dsads',
+          avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+          time: '2020-10-22 22:14:17',
+          likedCount: 100,
+          reply: [
+            {
+              text: ['123123hhh'],
+              userName: 'dsads',
+              avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+              time: '2020-10-22 22:14:17',
+              likedCount: 100,
+            },
+            {
+              text: ['123123hhh'],
+              userName: 'dsads',
+              avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+              time: '2020-10-22 22:14:17',
+              likedCount: 100,
+            },
+          ],
+        }, {
+          text: ['123123hhh'],
+          userName: 'dsads',
+          avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+          time: '2020-10-22 22:14:17',
+          likedCount: 100,
+          reply: [],
+        }, {
+          text: ['123123hhh'],
+          userName: 'dsads',
+          avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+          time: '2020-10-22 22:14:17',
+          likedCount: 100,
+          reply: [],
+        }, {
+          text: ['123123hhh'],
+          userName: 'dsads',
+          avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+          time: '2020-10-22 22:14:17',
+          likedCount: 100,
+          reply: [],
+        }, {
+          text: ['123123hhh'],
+          userName: 'dsads',
+          avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+          time: '2020-10-22 22:14:17',
+          likedCount: 100,
+          reply: [],
+        }, {
+          text: ['123123hhh'],
+          userName: 'dsads',
+          avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+          time: '2020-10-22 22:14:17',
+          likedCount: 100,
+          reply: [],
+        }, {
+          text: ['123123hhh'],
+          userName: 'dsads',
+          avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+          time: '2020-10-22 22:14:17',
+          likedCount: 100,
+          reply: [],
+        }],
+        hotComments: [{
+          text: ['123123hhh'],
+          userName: 'dsads',
+          avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+          time: '2020-10-22 22:14:17',
+          likedCount: 100,
+          reply: [
+            {
+              text: ['123123hhh'],
+              userName: 'dsads',
+              avatar: 'http://img4.kuwo.cn/star/userhead/39/52/1602393411654_512039239s.jpg',
+              time: '2020-10-22 22:14:17',
+              likedCount: 100,
+            },
+          ],
+        }],
+      },
     }
   },
   mounted() {
@@ -224,6 +339,11 @@ export default {
   computed: {
     ...mapGetters(['setting']),
     ...mapGetters('player', ['isShowPlayerDetail']),
+    title() {
+      return this.musicInfo.name
+        ? this.setting.download.fileName.replace('歌名', this.musicInfo.name).replace('歌手', this.musicInfo.singer)
+        : '^-^'
+    },
   },
   methods: {
     ...mapMutations('player', [
@@ -324,6 +444,9 @@ export default {
     },
     close() {
       window.eventHub.$emit(eventBaseName.close)
+    },
+    handleToggleCommentPage(page) {
+
     },
   },
 }
@@ -452,33 +575,56 @@ export default {
   overflow: hidden;
   display: flex;
   padding: 0 30px;
+
+  &.showComment {
+    .left {
+      flex-basis: 18%;
+      .description p {
+        font-size: 12px;
+      }
+    }
+    .right {
+      flex-basis: 30%;
+      .lyric {
+        font-size: 13px;
+      }
+    }
+    .comment {
+      flex-basis: 50%;
+      opacity: 1;
+    }
+  }
 }
 .left {
-  flex: 0 0 40%;
-  overflow: hidden;
-}
-.info {
+  flex: auto;
   display: flex;
   flex-flow: column nowrap;
   align-items: center;
+  padding: 13px;
+  overflow: hidden;
+  transition: flex-basis @transition-theme;
+}
+
+.info {
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: flex-start;
+  max-width: 300px;
+
 }
 .img {
-  width: 300px;
-  height: 300px;
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-
-  img {
-    max-width: 100%;
-    max-height: 100%;
-    border: 5px solid @color-theme-hover;
-    // border-radius: @radius-border;
-    // border: 5px solid #fff;
-  }
+  max-width: 100%;
+  max-height: 100%;
+  min-width: 100%;
+  box-shadow: 0 0 4px @color-theme-hover;
+  border-radius: 6px;
+  opacity: .8;
+  // border: 5px solid @color-theme-hover;
+  // border-radius: @radius-border;
+  // border: 5px solid #fff;
 }
 .description {
-  width: 300px;
+  max-width: 300px;
   padding: 15px 0;
   overflow: hidden;
   p {
@@ -492,6 +638,8 @@ export default {
   flex: 0 0 60%;
   // padding: 0 30px;
   position: relative;
+  transition: flex-basis @transition-theme;
+
   &:before {
     position: absolute;
     top: 0;
@@ -526,7 +674,7 @@ export default {
     padding: 8px 0;
     line-height: 1.2;
     overflow-wrap: break-word;
-    transition: @transition-theme;
+    transition: @transition-theme !important;
     transition-property: color, font-size;
   }
 }
@@ -537,19 +685,48 @@ export default {
   color: @color-theme;
   font-size: 1.2em;
 }
+
+.comment {
+  flex: 0 0 0;
+  display: flex;
+  flex-flow: column nowrap;
+  transition: @transition-theme;
+  transition-property: flex-basis opacity;
+  overflow: hidden;
+  opacity: 0;
+
+}
+.commentHeader {
+  flex: none;
+  padding-bottom: 10px;
+  // border-bottom: 1px solid #eee;
+  h3 {
+    font-size: 14px;
+  }
+}
+.commentMain {
+  flex: auto;
+  padding-left: 15px;
+  padding-right: 10px;
+  background-color: @color-reply-floor;
+  border-radius: 4px;
+}
+.commentType {
+  padding: 10px 0;
+  font-size: 13px;
+  color: @color-theme_2-font;
+}
+.pagination {
+  padding: 10px 0;
+}
+
 .footer {
   flex: 0 0 100px;
   overflow: hidden;
   display: flex;
   align-items: center;
 }
-.left {
-  flex: auto;
-  display: flex;
-  flex-flow: column nowrap;
-  align-items: center;
-  padding-top: 13px;
-}
+
 .progress-container {
   width: 100%;
   position: relative;
@@ -686,9 +863,8 @@ each(@themes, {
       }
     }
     .img {
-      img {
-        border-color: ~'@{color-@{value}-theme-hover}';
-      }
+      box-shadow: 0 0 4px ~'@{color-@{value}-theme-hover}';
+      // border-color: ~'@{color-@{value}-theme-hover}';
     }
     .lrc-active {
       color: ~'@{color-@{value}-theme}';