+++ src/renderer/components/core/Toolbar.vue
@@ -6,14 +6,14 @@
       @event="handleEvent" :list="tipList" :visibleList="visibleList"
       v-model="searchText")
 
-    div(:class="$style.logo" v-if="setting.controlBtnPosition == 'left'") L X
+    div(:class="$style.logo" v-if="setting.controlBtnPosition == 'left'")    
     div(:class="$style.control" v-else)
-      button(type="button" :class="$style.min" :title="$t('core.toolbar.min')" @click="min")
+      button(type="button" :class="$style.min" :tips="$t('core.toolbar.min')" @click="min")
         svg(:class="$style.icon" version='1.1' xmlns='http://www.w3.org/2000/svg' xlink='http://www.w3.org/1999/xlink' width='100%' viewBox='0 0 24 24' space='preserve')
           use(xlink:href='#icon-window-minimize')
 
       //- button(type="button" :class="$style.max" @click="max")
-      button(type="button" :class="$style.close" :title="$t('core.toolbar.close')" @click="close")
+      button(type="button" :class="$style.close" :tips="$t('core.toolbar.close')" @click="close")
         svg(:class="$style.icon" version='1.1' xmlns='http://www.w3.org/2000/svg' xlink='http://www.w3.org/1999/xlink' width='100%' viewBox='0 0 24 24' space='preserve')
           use(xlink:href='#icon-window-close')
 </template>
@@ -183,6 +183,7 @@ export default {
   }
 
   button {
+    display: flex;
     position: relative;
     width: @control-btn-width;
     height: @control-btn-width;