+++ src/renderer/store/modules/list.js
@@ -28,6 +28,12 @@ const state = {
     list: [],
     location: 0,
   },
+  tempList: {
+    id: 'temp',
+    name: '临时列表',
+    list: [],
+    location: 0,
+  },
   userList: [],
 }
 
@@ -48,8 +54,8 @@ const actions = {
 // mitations
 const mutations = {
   initList(state, { defaultList, loveList, userList }) {
-    if (defaultList != null) state.defaultList.list = defaultList.list
-    if (loveList != null) state.loveList.list = loveList.list
+    if (defaultList != null) Object.assign(state.defaultList, { list: defaultList.list, location: defaultList.location })
+    if (loveList != null) Object.assign(state.loveList, { list: loveList.list, location: loveList.location })
     if (userList != null) state.userList = userList
     allListInit(state.defaultList, state.loveList, state.userList)
     state.isInitedList = true
@@ -140,15 +146,19 @@ const mutations = {
     if (!targetList) return
     Object.assign(targetList.list[index], data)
   },
-  createUserList(state, name) {
-    let newList = {
-      name,
-      id: `userlist_${Date.now()}`,
-      list: [],
-      location: 0,
+  createUserList(state, { name, id = `userlist_${Date.now()}`, list = [] }) {
+    let newList = state.userList.find(item => item.id === id)
+    if (!newList) {
+      newList = {
+        name,
+        id,
+        list: [],
+        location: 0,
+      }
+      state.userList.push(newList)
+      allListUpdate(newList)
     }
-    state.userList.push(newList)
-    allListUpdate(newList)
+    this.commit('list/listAddMultiple', { id, list })
   },
   removeUserList(state, index) {
     let list = state.userList.splice(index, 1)[0]